Machine Learning
Using devices such as Jawbone Up, Nike FuelBand, and Fitbit it is now
possible to collect a large amount of data about personal activity
relatively inexpensively. These type of devices are part of the
quantified self movement – a group of enthusiasts who take measurements
about themselves regularly to improve their health, to find patterns in
their behavior, or because they are tech geeks. One thing that people
regularly do is quantify how much of a particular activity they do, but
they rarely quantify how well they do it. In this project, your goal
will be to use data from accelerometers on the belt, forearm, arm, and
dumbell of 6 participants. They were asked to perform barbell lifts
correctly and incorrectly in 5 different ways.  The goal of this
project is to predict the manner in which they did the exercise. This
is the "classe" variable in the training set. We should create a report
describing how we built your model, how we used cross validation, what
we think the expected out of sample error is, and why we made the
choices we did. We will also use your prediction model to predict 20
different test cases.